requirement.txt
transformers==4.31.0
datasets==2.13.1
torch>=1.13.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.4
tqdm==4.65.0
